Move Minitest overrides into separate file
